MGDCTRS-1729: connector still running even if internal KafkaConsumer thread was shut down
The default streategy for the KafkaConsumer is to delegate the error
handling to the Camel's error handler in case of an error while polling
kafka for messages and it then it should continue polling for messages
however, in case of authentication errors, the consumer is stopped and
no more attempts are made.

Because the message goes to the error handler (i.e. to a dead letter
queue), then the error is swallowed and the conenctor is not reported as
failed.

This commit changes the default behavior of the the KafkaConsumer to
retry to connect at least once which would also report the conenctor
as failed.

Signed-off-by: Luca Burgazzoli <lburgazzoli@gmail.com>